restructuring arts and newsletter files
