--- pages/newsletter/index.js
@@ -1,11 +1,0 @@
-import NewsletterCardGallery from "@/components/newsletter/NewsCard";
-import NewsHeaderCarousel from "./headerCarousel";
-
-export default function Art() {
-  return (
-    <>
-      <NewsHeaderCarousel />
-      <NewsletterCardGallery />
-    </>
-  );
-}
