+++ styles/art/Arts.module.css
@@ -1,0 +1,141 @@
+/* HEADER */
+
+.art-header {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 35vh;
+}
+
+.container {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  opacity: 0;
+  /* For fade-in effect */
+  transform: scale(0.95);
+  /* For initial scale animation */
+  transition: opacity 0.5s ease-out, transform 0.5s ease-out;
+}
+
+.fadeIn {
+  opacity: 1;
+  /* Fully visible */
+  transform: scale(1);
+  /* Back to original size */
+}
+
+.slideUp {
+  animation: slide-up 0.8s ease-out;
+  /* Slide up effect */
+}
+
+.header-content {
+  color: white;
+  text-align: center;
+  width: 95%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.title {
+  font-size: 1.25rem;
+  font-weight: 700;
+}
+
+.subtitle {
+  color: #d9d9d9;
+  font-size: 0.8rem;
+}
+
+.cta-button {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 0.5rem;
+  background-color: #e43332;
+  width: 200px;
+  height: 40px;
+  border-radius: 41px;
+}
+
+.cta-button:hover {
+  cursor: pointer;
+}
+
+.cta-text-mobile {
+  font-weight: 700;
+  font-size: 1rem;
+}
+
+.cta-text-desktop {
+  display: none;
+  font-weight: 700;
+  font-size: 1rem;
+}
+
+@media (min-width: 640px) {
+  .title {
+    font-size: 1.875rem;
+  }
+
+  .subtitle {
+    font-size: 0.875rem;
+  }
+
+  .cta-button {
+    height: 55px;
+  }
+
+  .cta-text-mobile {
+    display: none;
+  }
+
+  .cta-text-desktop {
+    display: block;
+  }
+}
+
+@media (min-width: 768px) {
+  .header-content {
+    width: 70%;
+  }
+
+  .cta-button {
+    width: 300px;
+    gap: 0.75rem;
+  }
+}
+
+@media (min-width: 1024px) {
+  .header-content {
+    width: 35%;
+  }
+
+  .cta-button {
+    width: 366px;
+    gap: 2rem;
+  }
+}
+
+@keyframes slide-up {
+  from {
+    transform: translateY(20px);
+    /* Start from below */
+    opacity: 0;
+    /* Invisible */
+  }
+
+  to {
+    transform: translateY(0);
+    /* End at original position */
+    opacity: 1;
+    /* Fully visible */
+  }
+}
